bash scripts and other files that help automate BBB testing

+++ tests/acceptance/test_update.py
@@ -22,6 +22,13 @@ import subprocess
 from common import *
 
 
+class Helpers:
+    @staticmethod
+    def upload_to_s3():
+        subprocess.Popen(["s3cmd", "--follow-symlinks", "put", "image.dat", "s3://yocto-builds/tmp/"]).wait()
+        subprocess.Popen(["s3cmd", "setacl", "s3://yocto-builds/tmp/image.dat", "--acl-public"]).wait()
+
+
 @pytest.mark.usefixtures("qemu_running", "no_image_file", "setup_bbb")
 class TestUpdates:
 
@@ -128,6 +135,8 @@ class TestUpdates:
 
     def test_network_based_image_update(self):
         http_server_location = pytest.config.getoption("--http-server")
+        bbb = pytest.config.getoption("--bbb")
+
         if not env.host_string:
             # This means we are not inside execute(). Recurse into it!
             execute(self.test_network_based_image_update)
@@ -136,13 +145,18 @@ class TestUpdates:
         output = run("mount")
         (active_before, passive_before) = determine_active_passive_part(output)
 
-        http_server = subprocess.Popen(["python", "-m", "SimpleHTTPServer"])
-        assert(http_server)
+        if bbb:
+            Helpers.upload_to_s3()
+            http_server_location = "s3-eu-west-1.amazonaws.com/yocto-builds/tmp"
+        else:
+            http_server = subprocess.Popen(["python", "-m", "SimpleHTTPServer"])
+            assert(http_server)
 
         try:
             sudo("mender -rootfs http://%s/image.dat" % (http_server_location))
         finally:
-            http_server.terminate()
+            if not bbb:
+                http_server.terminate()
 
         output = run("fw_printenv bootcount")
         assert(output == "bootcount=0")
